pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "NeoPortfolio"
version = "1.3.0"
description = "Portfolio Selection, Optimization, and backtesting library with sentiment analysis and ML features"
authors = [
    {name = "Ege Güney Kıymaç", email = "guneykiymac@gmail.com"},
    {name = "Cem Köymen", email = "cemkoymen5@gmail.com"}
]

readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "beautifulsoup4>=4.11.2",
    "ipython>=8.14.0",
    "matplotlib>=3.7.1",
    "newsapi-python>=0.2.7",
    "numpy>=1.24.3",
    "pandas>=2.0.2",
    "plotly>=5.24.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "requests-cache>=1.2.1",
    "scikit_learn>=1.2.2",
    "scipy>=1.15.0",
    "torch>=2.5.1",
    "tqdm>=4.65.0",
    "transformers>=4.47.1",
    "yfinance>=0.2.33"
]

[project.urls]
homepage = "https://github.com/GongJr0/NeoPortfolio"
documentation = "https://gongjr0.github.io/NeoPortfolio"
